Support MetaGraph in VM
